+++ libs/wingc/src/closure_transform.rs
@@ -370,29 +370,3 @@ impl<'a> Fold for RenameThisTransformer<'a> {
 		}
 	}
 }
-
-pub struct TypeReferenceTransformer<'a> {
-	pub types: &'a mut crate::type_check::Types,
-}
-
-impl<'a> Fold for TypeReferenceTransformer<'a> {
-	fn fold_scope(&mut self, node: Scope) -> Scope {
-		if self.types.type_expressions.is_empty() {
-			return node;
-		} else {
-			fold::fold_scope(self, node)
-		}
-	}
-	fn fold_reference(&mut self, node: Reference) -> Reference {
-		match node {
-			Reference::InstanceMember { ref object, .. } => {
-				if let Some(new_ref) = self.types.type_expressions.remove(&object.id) {
-					new_ref
-				} else {
-					fold::fold_reference(self, node)
-				}
-			}
-			Reference::Identifier(..) | Reference::TypeMember { .. } => fold::fold_reference(self, node),
-		}
-	}
-}
